+++ banhammer/subreddit.py
@@ -110,8 +110,16 @@ class Subreddit:
             if reaction.emoji == emoji:
                 return reaction
 
-    def get_new(self):
-        if not self.stream_new:
+    def ignore_old(self):
+        for p in self.get_new(True): break
+        for p in self.get_comments(True): break
+        for p in self.get_reports(True): break
+        for p in self.get_mail(True): break
+        for p in self.get_queue(True): break
+        for p in self.get_mod_actions(override=True): break
+
+    def get_new(self, override=False):
+        if not self.stream_new and not override:
             return list()
         path = "files/{}_new.txt".format(self.subreddit.id)
         ids = list()
@@ -125,8 +133,8 @@ class Subreddit:
             item.save(path)
             yield item
 
-    def get_comments(self):
-        if not self.stream_comments:
+    def get_comments(self, override=False):
+        if not self.stream_comments and not override:
             return list()
         path = "files/{}_comments.txt".format(self.subreddit.id)
         ids = list()
@@ -134,14 +142,14 @@ class Subreddit:
             with open(path) as f:
                 ids = f.read().splitlines()
         for comment in self.subreddit.comments(limit=250):
-            if submission.id in ids:
+            if comment.id in ids:
                 break
             item = RedditItem(comment, self, "new")
             item.save(path)
             yield item
 
-    def get_reports(self):
-        if not self.stream_reports:
+    def get_reports(self, override=False):
+        if not self.stream_reports and not override:
             return list()
         path = "files/{}_reports.txt".format(self.subreddit.id)
         ids = list()
@@ -155,8 +163,8 @@ class Subreddit:
             item.save(path)
             yield item
 
-    def get_mail(self):
-        if not self.stream_mail:
+    def get_mail(self, override=False):
+        if not self.stream_mail and not override:
             return list()
         path = "files/{}_mail.txt".format(self.subreddit.id)
         ids = list()
@@ -172,8 +180,8 @@ class Subreddit:
                 item.save(path)
                 yield item
 
-    def get_queue(self):
-        if not self.stream_queue:
+    def get_queue(self, override=False):
+        if not self.stream_queue and not override:
             return list()
         path = "files/{}_queue.txt".format(self.subreddit.id)
         ids = list()
@@ -187,9 +195,9 @@ class Subreddit:
             item.save(path)
             yield item
 
-    def get_mod_actions(self, mods):
+    def get_mod_actions(self, mods=list(), override=False):
         mods = [m.lower() for m in mods]
-        if not self.stream_mod_actions:
+        if not self.stream_mod_actions and not override:
             return list()
         path = "files/{}_actions.txt".format(self.subreddit.id)
         ids = list()